Droit XML (re)moved to legacy
